al dia con master

+++ app/Http/Resources/ProductCollection.php
@@ -9,19 +9,19 @@ class ProductCollection extends ResourceCollection
     public function toArray($request)
     {
         return [
-            'data' => $this->collection->map(function($data) {
+            'data' => $this->collection->map(function ($data) {
                 return [
-                    'id' => (integer) $data->id,
+                    'id' => (int) $data->id,
                     'name' => $data->getTranslation('name'),
                     'slug' => $data->slug,
                     'thumbnail_image' => api_asset($data->thumbnail_img),
-                    'base_price' => (double) $data->lowest_price,
-                    'base_discounted_price' => (double) product_discounted_base_price($data),
+                    'base_price' => (float) $data->lowest_price,
+                    'base_discounted_price' => (float) product_discounted_base_price($data),
                     'stock' => $data->stock,
                     'unit' => $data->getTranslation('unit'),
                     'min_qty' => $data->min_qty,
                     'max_qty' => $data->max_qty,
-                    'rating' => (double) $data->rating,
+                    'rating' => (float) $data->rating,
                     'is_variant' => (int) $data->is_variant,
                     'variations' => $data->variations,
                     'description' => $data->description,
